feat!: add ip endpoint
Fixes: https://github.com/neeythann/ip-location-rs/issues/33
Fixes: https://github.com/neeythann/ip-location-rs/issues/29

This commit fixes the failing mock tests in
https://github.com/neeythann/ip-location-rs/issues/29 by extracting it
as a resource name endpoint instead of a query parameter. Also added
tests for new said api endpoint.

This is also a breaking change since this commit changes the
functionality of the index endpoint. I will be skipping v0.1.6 without
any backport support on the previous commits after v0.1.5.